clean up naming and make tests more explicit
