+++ test/04-compiler-test.js
